+++ CHANGELOG.md
@@ -5,8 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-### Fixedd
+### Fixed
 - Fixed `Lexer is not defined` error when parsing a hex string #215
+- Added an error when multiple headers are added to a table #213
 
 ## [2.3.8] - 2020-05-22
 ### Fixed